Server session actions cleanup
* Move actions into a seperate server-session-actions file (this fixes the
  last of our input cycle problems)
* Remove login/logout "tasks", just dispatch these redux actions where
  needed
