+++ src/editor/user-tasks/task-definitions.js
@@ -57,20 +57,6 @@ tasks.evaluateChunkAndSelectBelow = new UserTask({
   }
 });
 
-tasks.loginGithub = new UserTask({
-  title: "Login using GitHub",
-  callback(loginSuccess = undefined) {
-    dispatcher.login(loginSuccess);
-  }
-});
-
-tasks.logoutGithub = new UserTask({
-  title: "Logout",
-  callback() {
-    dispatcher.logout();
-  }
-});
-
 tasks.toggleWrapInEditors = new UserTask({
   title: "Toggle wrapping in editors",
   displayKeybinding: "Alt+w", // '\u2193',
